Create nvd-cve-json-polling.py
This script will check for new CVEs published since the previous check.

On each iteration, the file 'last_fetch.txt' will be updated with a timestamp. On the next iteration, that timestamp will be used in the API request as 'pubStartDate', with the current time used as the 'pubEndDate'. This will create a delta to fetch all published CVEs between those times.

Output is logged as JSON in the 'cve_log.log file', with debug logging logged to 'output.log'.

log file rotation should occur at 5MB file size.